API

Types

BoostFractor.Modes

BoostFractor.SetupBoundaries

Functions

BoostFractor.SeedSetupBoundaries

BoostFractor.axion_S_factor

BoostFractor.axion_contrib

BoostFractor.axion_induced_modes

BoostFractor.calc_propagation_matrices

BoostFractor.cheerleader

BoostFractor.dance_intro

BoostFractor.dancer

BoostFractor.e_field_dimensions

BoostFractor.field2modes

BoostFractor.gauss_profile

BoostFractor.get_boundary_matrix

BoostFractor.get_kspace_coords

BoostFractor.index

BoostFractor.mode

BoostFractor.modes2field

BoostFractor.propagation_matrix

BoostFractor.propagator

BoostFractor.propagator1D

BoostFractor.propagatorMomentumSpace

BoostFractor.propagatorNoTilts

BoostFractor.transformer

BoostFractor.transformer_trace_back

Documentation

BoostFractor.Modes — Type

Modes

Doc: TODO

BoostFractor.SetupBoundaries — Type

Summary

mutable struct SetupBoundaries <: Any

Define properties of dielectric boundaries. Coordinate system?

Fields:

distance::Array{Float64,1} > 0: Distance in z direction to boundary

r::Array{Complex{Float64},1} [0, 1]: Boundary reflection coefficient for right-propagating wave

eps::Array{Complex{Float64},1}: Dielectric permittivity to the right of each boundary"

relative_tilt_x > 0: Tilt in x direction [rad?]

relative_tilt_y > 0: Tilt in y direction [rad?]

relative_surfaces::Array{Complex{Float64},3} ?: Surface roughness. z offset (1st dim) at x,y (2nd, 3rd dims)

BoostFractor.SeedSetupBoundaries — Method

SeedSetupBoundaries(diskno=3)

Initialize mutable struct SetupBoundaries with sensible values.

Arguments

diskno::Int > 0: Number of dielectric discs

BoostFractor.axion_S_factor — Method

Prefactor in Theoretical Foundations (4.7c).

BoostFractor.axion_contrib — Method

Calculates one summand of the term (M[2,1]+M[1,1]) E0 = sum{s=1...m} (Ts^m[2,1]+Ts^m[1,1]) E0 as in equation 4.14a

BoostFractor.axion_induced_modes — Method

Calculate the vector of mode-coefficients that describe the uniform axion-induced field and that is normalized to power 1.

Implements Knirck 6.17.

BoostFractor.calc_propagation_matrices — Method

Pre-calculates all the propagation matrices. Useful, if they should be altered later (e.g. take out modes, add some additional mixing, etc.)

BoostFractor.cheerleader — Method

cheerleader(amin, nmax, bdry::SetupBoundaries; f=10.0e9, prop=propagator, emit=nothing, reflect=nothing, Xset=X, Yset=Y, diskR=0.1, returnboth=false)

New Recursive Fourier Propagation implementation.

Arguments:

amin: Mimum (local) amplitude of a field, in order to be propagated

nmax: Maximum number of beam iteration steps, directly equivalent to how many boundaries a beam 'sees' at most

bdry::SetupBoundaries: Properties of dielectric boundaries

f::Float64 > 0: Frequency of EM radiation

prop: Propagator Function to use. Standard is propagator().

emit: Explicitly set the axion-induced fields emitted from each boundary (to the left and to the right). If $nothing$ fields are initialized according to uniform, homogeneous external B-field with zero axion velocity.

reflect: If nothing (standar value), the axion-induced signal is computed. If set, this field defines a beam, for which the reflected beam will be calculated

Xset and Yset: Explicitly set the coordinate system for the fields

diskR: Radius of dielectric disk

returnboth::Bool: If true cheerleader returns fields leaving on left and right. If false only returns fields leaving on right.

See dancer for old version.

BoostFractor.dance_intro — Method

dance_intro(bdry::SetupBoundaries, X, Y; bfield=nothing, velocity_x=0, f=10e9,diskR=0.1)

Initialize EM fields. Can include velocity effects.

BoostFractor.dancer — Method

Propagates the fields through the system

amin: Mimum (local) amplitude of a field, in order to be propagated

nmax: Maximum number of beam iteration steps, directly equivalent to how many boundaries a beam 'sees' at most

bdry: SetupBoundaries-Objekt, containing all relevant geometrical information (disk positions, epsilon, etc)

f: Frequency

prop: Propagator Function to use. Standard is propagator()

reflect: If nothing (standar value), the axion-induced signal is computed. If set, this field defines a beam, for which the reflected beam will be calculated

Xset, Yset: Explicitly set the coordinate system for the fields

returnsum: If false, the out-propagating contributions after each iteration will be returned, without summing.

immediatesum: If false, the out-propagating contributions will be saved and summed up at the end.

BoostFractor.e_field_dimensions — Method

Gives out E-field dimensions in a more readable way. Output should be either 1 or 3.

BoostFractor.field2modes — Method

Get the mode coefficients for a given field distribution E(x,y)

BoostFractor.gauss_profile — Method

gauss_profile(coords::CoordinateSystem; z = 0.000000001, omega0 = 0.1, f = 10e9)

Compute normalized 2D gaussian beam profile at given z position (waist at z = 0).

Arguments

coords: contains X,Y and Kx,Ky coordinates

z: Distance from beam waist

omega0 > 0: Beam waist radius

f > 0: Frequency of monochromatic EM radiation

BoostFractor.get_boundary_matrix — Method

Calculate Transfer matrix like in 4.9.

BoostFractor.get_kspace_coords — Method

get_kspace_coords(RealSpaceCoords)

Calculate k space coordinate system from real space one. Helper function for SeedCoordinateSystem.

BoostFractor.index — Method

index(modes::Modes, k::Int64)

Indexing function to get sub-matrices. (Compare Knirck 6.18.)

BoostFractor.mode — Method

mode(m,l, coords::CoordinateSystem; diskR=0.15, k0=2pi/0.03)

Calculate the transverse k and field distribution for a mode. Implements Knirck 6.12.

BoostFractor.modes2field — Method

Get the field distribution E(x,y) for a given vector of mode coefficients

BoostFractor.propagation_matrix — Method

BoostFractor.propagator — Method

propagator(E0, dz, diskR, eps, tilt_x, tilt_y, surface, lambda)

Do the FFT of E0 on a disk, propagate the beam a given distance and do the iFFT. Note that this method is in-place. If it should be called more than one time on the same fields, use propagator(copy(E0), ...).

Assume: Tilt is small, additional phase is obtained by propagating all fields just with k0 to the tilted surface (only valid if diffraction effects are small).

Arguments

E0::Array{Float64,2}: Electric field before propagation

dz: Distance propagated in z direction

diskR: Radius of discs

eps: Dielectric permittivity

tilt_x: Disc tilt in x direction

tilt_y: Disc tilt in y direction

surface: Surface roughness of disc (only at end of propagation)

lambda: Wavelength of electric field

See also: propagatorMomentumSpace

BoostFractor.propagator1D — Method

propagator1D(E0, dz, diskR, eps, tilt_x, tilt_y, surface, lambda)

This propagator just does the phase propagation. Go to propagator for documentation. 3D arguments to be compatible with other propagators.

BoostFractor.propagatorMomentumSpace — Method

propagatorMomentumSpace(E0, dz, diskR, eps, tilt_x, tilt_y, surface, lambda)

Propagator that assumes E0 is already in momentum space. Mix between propagator and propagatorNoTilts. Go to propagator for documentation.

BoostFractor.propagatorNoTilts — Method

propagatorNoTilts(E0, dz, diskR, eps, tilt_x, tilt_y, surface, lambda)

Wrapped by propagator. Go there for documentation. Tilt arguments to be compatible with other propagators.

BoostFractor.transformer — Method

Transformer Algorithm using Transfer Matrices and Modes to do the 3D Calculation.

BoostFractor.transformer_trace_back — Method

Traces back the field of a solution for the reflectivity in the system. Same arguments as transformer(), but first two arguments are the reflected beam solution and the input beam. Returns an array with the left- and rightgoing field amplitudes in each region.

Theoreticallly works in 3D as well, however be careful when plotting the output! Simple multiplication with e^{ikz} doesnt work anymore for 3D. Need some sort of Fourier propagation to specific points between disks to be able to plot result.